+++ src/Microsoft.AspNet.Authorization/AuthorizationServiceExtensions.cs
@@ -20,12 +20,6 @@ namespace Microsoft.AspNet.Authorization
         /// <returns><value>true</value> when the user fulfills the policy, <value>false</value> otherwise.</returns>
         public static Task<bool> AuthorizeAsync([NotNull] this IAuthorizationService service, ClaimsPrincipal user, object resource, [NotNull] AuthorizationPolicy policy)
         {
-            // TODO RENABLE
-            //if (policy.ActiveAuthenticationSchemes != null && policy.ActiveAuthenticationSchemes.Any() && user != null)
-            //{
-            //    // Filter the user to only contain the active authentication types
-            //    user = new ClaimsPrincipal(user.Identities.Where(i => policy.ActiveAuthenticationSchemes.Contains(i.AuthenticationScheme)));
-            //}
             return service.AuthorizeAsync(user, resource, policy.Requirements.ToArray());
         }
 
@@ -39,12 +33,6 @@ namespace Microsoft.AspNet.Authorization
         /// <returns><value>true</value> when the user fulfills the policy, <value>false</value> otherwise.</returns>
         public static bool Authorize([NotNull] this IAuthorizationService service, ClaimsPrincipal user, object resource, [NotNull] AuthorizationPolicy policy)
         {
-            // TODO: REeanble
-            //if (policy.ActiveAuthenticationSchemes != null && policy.ActiveAuthenticationSchemes.Any() && user != null)
-            //{
-            //    // Filter the user to only contain the active authentication types
-            //    user = new ClaimsPrincipal(user.Identities.Where(i => policy.ActiveAuthenticationSchemes.Contains(i.AuthenticationScheme)));
-            //}
             return service.Authorize(user, resource, policy.Requirements.ToArray());
         }
