--- src/Microsoft.AspNet.Authorization/ClaimsAuthorizationHandler.cs
@@ -1,32 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Linq;
-
-namespace Microsoft.AspNet.Authorization
-{
-    public class ClaimsAuthorizationHandler : AuthorizationHandler<ClaimsAuthorizationRequirement>
-    {
-        public override void Handle(AuthorizationContext context, ClaimsAuthorizationRequirement requirement)
-        {
-            if (context.User != null)
-            {
-                bool found = false;
-                if (requirement.AllowedValues == null || !requirement.AllowedValues.Any())
-                {
-                    found = context.User.Claims.Any(c => string.Equals(c.Type, requirement.ClaimType, StringComparison.OrdinalIgnoreCase));
-                }
-                else
-                {
-                    found = context.User.Claims.Any(c => string.Equals(c.Type, requirement.ClaimType, StringComparison.OrdinalIgnoreCase)
-                                                        && requirement.AllowedValues.Contains(c.Value, StringComparer.Ordinal));
-                }
-                if (found)
-                {
-                    context.Succeed(requirement);
-                }
-            }
-        }
-    }
-}
