--- src/Microsoft.AspNet.Authorization/DenyAnonymousAuthorizationHandler.cs
@@ -1,22 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Linq;
-
-namespace Microsoft.AspNet.Authorization
-{
-    public class DenyAnonymousAuthorizationHandler : AuthorizationHandler<DenyAnonymousAuthorizationRequirement>
-    {
-        public override void Handle(AuthorizationContext context, DenyAnonymousAuthorizationRequirement requirement)
-        {
-            var user = context.User;
-            var userIsAnonymous =
-                user?.Identity == null ||
-                !user.Identities.Any(i => i.IsAuthenticated);
-            if (!userIsAnonymous)
-            {
-                context.Succeed(requirement);
-            }
-        }
-    }
-}
