+++ src/Microsoft.AspNet.Authorization/DenyAnonymousAuthorizationRequirement.cs
@@ -1,9 +1,22 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNet.Authorization;
+using System.Linq;
 
 namespace Microsoft.AspNet.Authorization
 {
-    public class DenyAnonymousAuthorizationRequirement : IAuthorizationRequirement { }
+    public class DenyAnonymousAuthorizationRequirement : AuthorizationHandler<DenyAnonymousAuthorizationRequirement>, IAuthorizationRequirement
+    {
+        public override void Handle(AuthorizationContext context, DenyAnonymousAuthorizationRequirement requirement)
+        {
+            var user = context.User;
+            var userIsAnonymous =
+                user?.Identity == null ||
+                !user.Identities.Any(i => i.IsAuthenticated);
+            if (!userIsAnonymous)
+            {
+                context.Succeed(requirement);
+            }
+        }
+    }
 }
