+++ src/Microsoft.AspNet.Authorization/Resources.resx
@@ -117,16 +117,7 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Exception_DefaultDpapiRequiresAppNameKey" xml:space="preserve">
-    <value>The default data protection provider may only be used when the IApplicationBuilder.Properties contains an appropriate 'host.AppName' key.</value>
-  </data>
-  <data name="Exception_UnhookAuthenticationStateType" xml:space="preserve">
-    <value>The state passed to UnhookAuthentication may only be the return value from HookAuthentication.</value>
-  </data>
-  <data name="Exception_AuthenticationTokenDoesNotProvideSyncMethods" xml:space="preserve">
-    <value>The AuthenticationTokenProvider's required synchronous events have not been registered.</value>
-  </data>
-  <data name="Exception_AuthorizationPolicyNotFound" xml:space="preserve">
-    <value>The AuthorizationPolicy named: '{0}' was not found.</value>
+  <data name="Exception_RoleRequirementEmpty" xml:space="preserve">
+    <value>At least one role must be specified.</value>
   </data>
 </root>
